Neurostack-RAG-FINAL/backend/requirements.txt
# --- Core Web/API Frameworks ---
fastapi
uvicorn
pydantic
bcrypt==4.0.1
python-jose
sqlalchemy

# --- Rag and llm dependencies
langchain # Core LangChain library
langchain-core
langchain-community # Includes support for FAISS, BM25, and other components
groq    # For calling the LLM
faiss-cpu    # Vector store (FAISS)
rank_bm25
sentence-transformers # Needed for HuggingFaceEmbeddings
python-dotenv 
# For reading the .env file (GROQ_API_KEY, etc.)
requests # For general HTTP calls (already in your list, but essential)
PyJWT # For validating JSON Web Tokens (JWT)
python-multipart # To handle form data (if needed, good practice)
pypdf

SQLAlchemy
